Rename the TeammateRevive.Debug namespace to TeammateRevice.Debugging
The Debug directory is generally ignored in git repositories so it is
better to change the name to something that will be ignored by mistake.
